+++ packages/tagscript-plugin-discord/src/lib/Transformer/index.ts
@@ -1,1 +1,7 @@
-export * from './DiscordJs';
+export * from './Base';
+export * from './Guild';
+export * from './GuildMember';
+export * from './GuildMember';
+export * from './GuildTextBasedChannel';
+export * from './Role';
+export * from './User';
